Add ability to /aim by right-clicking with a clock
