+++ src/main/java/net/tylers1066/movecraftcannons/MovecraftCannons.java
@@ -14,6 +14,7 @@ import net.countercraft.movecraft.craft.CraftType;
 import net.countercraft.movecraft.util.MathUtils;
 import net.tylers1066.movecraftcannons.aiming.AimingCommand;
 import net.countercraft.movecraft.util.hitboxes.HitBox;
+import net.tylers1066.movecraftcannons.aiming.AimingListener;
 import net.tylers1066.movecraftcannons.config.Config;
 import net.tylers1066.movecraftcannons.listener.*;
 import net.tylers1066.movecraftcannons.localisation.I18nSupport;
@@ -114,6 +115,8 @@ public final class MovecraftCannons extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new TranslationListener(), this);
         getServer().getPluginManager().registerEvents(new RotationListener(), this);
         getServer().getPluginManager().registerEvents(new CannonListener(), this);
+        getServer().getPluginManager().registerEvents(new AimingListener(), this);
+
         this.getCommand("aim").setExecutor(new AimingCommand());
     }
