+++ src/main/java/net/tylers1066/movecraftcannons/aiming/AimingListener.java
@@ -1,0 +1,33 @@
+package net.tylers1066.movecraftcannons.aiming;
+
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import org.bukkit.Bukkit;
+import org.bukkit.Material;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.Action;
+import org.bukkit.event.player.PlayerInteractEvent;
+
+public class AimingListener implements Listener {
+
+    @EventHandler(priority = EventPriority.MONITOR)
+    public void onInteract(PlayerInteractEvent event) {
+        if (event.getAction() != Action.RIGHT_CLICK_AIR && event.getAction() != Action.RIGHT_CLICK_BLOCK) {
+            return;
+        }
+
+        if (event.getItem() == null || event.getItem().getType() != Material.CLOCK) {
+            return;
+        }
+
+        Player player = event.getPlayer();
+        Craft craft = CraftManager.getInstance().getCraftByPlayer(player);
+        if (craft == null) {
+            return;
+        }
+        AimingUtils.aimCannonsOnCraft(craft, player);
+    }
+}
