+++ src/main/java/net/tylers1066/movecraftcannons/utils/MovecraftUtils.java
@@ -1,5 +1,7 @@
 package net.tylers1066.movecraftcannons.utils;
 
+import at.pavlov.cannons.cannon.Cannon;
+import at.pavlov.cannons.utils.CannonsUtil;
 import com.palmergames.bukkit.towny.TownyAPI;
 import com.palmergames.bukkit.towny.object.Resident;
 import com.palmergames.bukkit.towny.utils.CombatUtil;
@@ -8,8 +10,13 @@ import net.countercraft.movecraft.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
 import net.countercraft.movecraft.craft.PlayerCraft;
 import net.countercraft.movecraft.util.MathUtils;
+import net.tylers1066.movecraftcannons.aiming.AimingCommand;
+import net.tylers1066.movecraftcannons.listener.DetectionListener;
+import net.tylers1066.movecraftcannons.localisation.I18nSupport;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
+
+import java.util.Set;
 
 public class MovecraftUtils {
 
@@ -36,5 +43,6 @@ public class MovecraftUtils {
         }
         return null;
     }
+
 
 }
